add generalizability to more curve pools
